refactor: 디렉토리 이름 변경 및 팔로우 로직 변경
