+++ src/main/java/com/example/newsfeed/follow/controller/FollowController.java
@@ -1,0 +1,39 @@
+package com.example.newsfeed.follow.controller;
+
+import com.example.newsfeed.global.common.Const.SessionConst;
+import com.example.newsfeed.global.response.Response;
+import com.example.newsfeed.follow.application.service.FollowService;
+import jakarta.servlet.http.HttpSession;
+import lombok.RequiredArgsConstructor;
+import org.springframework.web.bind.annotation.*;
+
+import java.util.List;
+
+@RestController
+@RequestMapping("/users/{userId}")
+@RequiredArgsConstructor
+public class FollowController {
+    private final FollowService followService;
+
+    @PostMapping("follows/{targetUserId}")
+    public Response<String> follow(
+            HttpSession session,
+            @PathVariable Long targetUserId
+    ) {
+        Long sessionUserId = (Long) session.getAttribute(SessionConst.LOGIN_USER);
+        boolean isFollowing = followService.toggleFollow(sessionUserId, targetUserId);
+        return Response.of(isFollowing ? "팔로우" : "언팔로우");
+    }
+
+    @GetMapping("followings/count")
+    public Response<List<Long>> findFollowingIdsByUserId(@PathVariable Long userId) {
+        List<Long> followingIds = followService.findFollowingIdsByUserId(userId);
+        return Response.of(followingIds);
+    }
+
+    @GetMapping("followers/count")
+    public Response<List<Long>> findFollowerIdsByUserId(@PathVariable Long userId) {
+        List<Long> followingIds = followService.findFollowerIdsByUserId(userId);
+        return Response.of(followingIds);
+    }
+}
