+++ src/main/java/com/example/newsfeed/global/common/exception/ErrorCode.java
@@ -16,9 +16,8 @@ public enum ErrorCode {
     FILE_NOT_FOUND("FileNotFound", "파일을 찾을 수 없습니다."),
     PASSWORD_SAME_AS_OLD("PasswordSameAsOld", "이전 패스워드와 동일할 수 없습니다."),
     FEED_NOT_FOUND("FeedNotFound", "해당 피드를 찾을 수 없습니다."),
-    NO_PERMISSION_AT_UPDATE_FEED("NoPermissionAtUpdateFeed", "해당 피드의 작성자만 수정할 수 있습니다.");
-    CANNOT_FOLLOW_SELF("CannotFollowSelf","자기 자신을 팔로우할 수 없습니다."),
-    FEED_NOT_FOUND("FeedNotFound", "해당 피드를 찾을 수 없습니다.");
+    NO_PERMISSION_AT_UPDATE_FEED("NoPermissionAtUpdateFeed", "해당 피드의 작성자만 수정할 수 있습니다."),
+    CANNOT_FOLLOW_SELF("CannotFollowSelf","자기 자신을 팔로우할 수 없습니다.");
 
     private final String code;
     private final String message;
